Merge master into production-v4

+++ app/Http/Controllers/AdminController.php
@@ -2,7 +2,6 @@
 
 namespace App\Http\Controllers;
 
-use App\Agave;
 use App\CachedSample;
 use App\Download;
 use App\FieldName;
@@ -26,6 +25,12 @@ class AdminController extends Controller
 {
     public function getQueues()
     {
+        // Check to see if user is Admin, if not return unautorized message.
+        $user = User::where('username', auth()->user()->username)->first();
+        if ($user == null || ! $user->isAdmin()) {
+            abort(401, 'Not authorized.');
+        }
+
         $jobs = [];
         $jobs['admin'] = LocalJob::findLast('admin');
         $jobs['short-downloads'] = LocalJob::findLast('short-downloads');
@@ -42,6 +47,12 @@ class AdminController extends Controller
 
     public function getDatabases()
     {
+        // Check to see if user is Admin, if not return unautorized message.
+        $user = User::where('username', auth()->user()->username)->first();
+        if ($user == null || ! $user->isAdmin()) {
+            abort(401, 'Not authorized.');
+        }
+
         $rs_list = RestService::findAvailable();
 
         $data = [];
@@ -53,6 +64,12 @@ class AdminController extends Controller
 
     public function getUpdateDatabase($id, $enabled)
     {
+        // Check to see if user is Admin, if not return unautorized message.
+        $user = User::where('username', auth()->user()->username)->first();
+        if ($user == null || ! $user->isAdmin()) {
+            abort(401, 'Not authorized.');
+        }
+
         $enabled = filter_var($enabled, FILTER_VALIDATE_BOOLEAN);
 
         $rs = RestService::find($id);
@@ -67,6 +84,12 @@ class AdminController extends Controller
 
     public function getDatabaseStats($id)
     {
+        // Check to see if user is Admin, if not return unautorized message.
+        $user = User::where('username', auth()->user()->username)->first();
+        if ($user == null || ! $user->isAdmin()) {
+            abort(401, 'Not authorized.');
+        }
+
         $response_list = RestService::samples([], 'titi', true, [$id]);
 
         $sample_list = [];
@@ -85,6 +108,12 @@ class AdminController extends Controller
 
     public function getNews()
     {
+        // Check to see if user is Admin, if not return unautorized message.
+        $user = User::where('username', auth()->user()->username)->first();
+        if ($user == null || ! $user->isAdmin()) {
+            abort(401, 'Not authorized.');
+        }
+
         $data = [];
         $data['news_list'] = News::orderBy('created_at', 'desc')->get();
         $data['notification'] = session()->get('notification');
@@ -94,6 +123,12 @@ class AdminController extends Controller
 
     public function getAddNews()
     {
+        // Check to see if user is Admin, if not return unautorized message.
+        $user = User::where('username', auth()->user()->username)->first();
+        if ($user == null || ! $user->isAdmin()) {
+            abort(401, 'Not authorized.');
+        }
+
         $data = [];
 
         return view('news/add', $data);
@@ -101,6 +136,12 @@ class AdminController extends Controller
 
     public function postAddNews(Request $request)
     {
+        // Check to see if user is Admin, if not return unautorized message.
+        $user = User::where('username', auth()->user()->username)->first();
+        if ($user == null || ! $user->isAdmin()) {
+            abort(401, 'Not authorized.');
+        }
+
         // validate form
         $rules = [
             'message' => 'required',
@@ -129,6 +170,12 @@ class AdminController extends Controller
 
     public function getEditNews($id)
     {
+        // Check to see if user is Admin, if not return unautorized message.
+        $user = User::where('username', auth()->user()->username)->first();
+        if ($user == null || ! $user->isAdmin()) {
+            abort(401, 'Not authorized.');
+        }
+
         $news = News::find($id);
 
         $data = [];
@@ -139,6 +186,12 @@ class AdminController extends Controller
 
     public function postEditNews(Request $request)
     {
+        // Check to see if user is Admin, if not return unautorized message.
+        $user = User::where('username', auth()->user()->username)->first();
+        if ($user == null || ! $user->isAdmin()) {
+            abort(401, 'Not authorized.');
+        }
+
         // validate form
         $rules = [
             'message' => 'required',
@@ -168,14 +221,26 @@ class AdminController extends Controller
 
     public function getDeleteNews($id)
     {
+        // Check to see if user is Admin, if not return unautorized message.
+        $user = User::where('username', auth()->user()->username)->first();
+        if ($user == null || ! $user->isAdmin()) {
+            abort(401, 'Not authorized.');
+        }
+
         $n = News::find($id);
         $n->delete();
 
-        return redirect('admin/users')->with('notification', 'News was successfully deleted.');
+        return redirect('admin/news')->with('notification', 'News was successfully deleted.');
     }
 
     public function getUsers($sort = 'created_at')
     {
+        // Check to see if user is Admin, if not return unautorized message.
+        $user = User::where('username', auth()->user()->username)->first();
+        if ($user == null || ! $user->isAdmin()) {
+            abort(401, 'Not authorized.');
+        }
+
         $l = User::orderByDesc('created_at')->get();
 
         $data = [];
@@ -187,6 +252,12 @@ class AdminController extends Controller
 
     public function getUsers2($sort = 'created_at')
     {
+        // Check to see if user is Admin, if not return unautorized message.
+        $user = User::where('username', auth()->user()->username)->first();
+        if ($user == null || ! $user->isAdmin()) {
+            abort(401, 'Not authorized.');
+        }
+
         $l = User::orderByDesc('created_at')->get();
 
         $data = [];
@@ -197,6 +268,12 @@ class AdminController extends Controller
 
     public function getAddUser()
     {
+        // Check to see if user is Admin, if not return unautorized message.
+        $user = User::where('username', auth()->user()->username)->first();
+        if ($user == null || ! $user->isAdmin()) {
+            abort(401, 'Not authorized.');
+        }
+
         $data = [];
 
         return view('user/add', $data);
@@ -204,6 +281,12 @@ class AdminController extends Controller
 
     public function postAddUser(Request $request)
     {
+        // Check to see if user is Admin, if not return unautorized message.
+        $user = User::where('username', auth()->user()->username)->first();
+        if ($user == null || ! $user->isAdmin()) {
+            abort(401, 'Not authorized.');
+        }
+
         // validate form
         $rules = [
             'first_name' => 'required',
@@ -245,6 +328,12 @@ class AdminController extends Controller
 
     public function getEditUser($id)
     {
+        // Check to see if user is Admin, if not return unautorized message.
+        $user = User::where('username', auth()->user()->username)->first();
+        if ($user == null || ! $user->isAdmin()) {
+            abort(401, 'Not authorized.');
+        }
+
         $user = User::find($id);
 
         $data = [];
@@ -260,6 +349,12 @@ class AdminController extends Controller
 
     public function postEditUser(Request $request)
     {
+        // Check to see if user is Admin, if not return unautorized message.
+        $user = User::where('username', auth()->user()->username)->first();
+        if ($user == null || ! $user->isAdmin()) {
+            abort(401, 'Not authorized.');
+        }
+
         // validate form
         $rules = [
             'id' => 'required',
@@ -291,19 +386,14 @@ class AdminController extends Controller
         return redirect('admin/users')->with('notification', 'Modifications for ' . $user->first_name . ' ' . $user->lastName . ' were successfully saved.');
     }
 
-    public function getDeleteUser($username)
-    {
-        // create Agave account
-        $agave = new Agave;
-        $token = $agave->getAdminToken();
-        $agave->deleteUser($token, $username);
-
-        return redirect('admin/users')->with('notification', 'User ' . $username . ' was successfully deleted.');
-    }
-
     public function getUpdateSampleCache()
     {
-        $username = auth()->user()->username;
+        // Check to see if user is Admin, if not return unautorized message.
+        $user = User::where('username', auth()->user()->username)->first();
+        if ($user == null || ! $user->isAdmin()) {
+            abort(401, 'Not authorized.');
+        }
+
         $n = CachedSample::cache();
 
         $message = "$n samples have been retrieved and cached.";
@@ -316,6 +406,12 @@ class AdminController extends Controller
 
     public function getUpdateSequenceCount($rest_service_id)
     {
+        // Check to see if user is Admin, if not return unautorized message.
+        $user = User::where('username', auth()->user()->username)->first();
+        if ($user == null || ! $user->isAdmin()) {
+            abort(401, 'Not authorized.');
+        }
+
         $rs = RestService::find($rest_service_id);
         $username = auth()->user()->username;
 
@@ -337,6 +433,12 @@ class AdminController extends Controller
 
     public function getUpdateCloneCount($rest_service_id)
     {
+        // Check to see if user is Admin, if not return unautorized message.
+        $user = User::where('username', auth()->user()->username)->first();
+        if ($user == null || ! $user->isAdmin()) {
+            abort(401, 'Not authorized.');
+        }
+
         $rs = RestService::find($rest_service_id);
         $username = auth()->user()->username;
 
@@ -358,6 +460,12 @@ class AdminController extends Controller
 
     public function getUpdateCellCount($rest_service_id)
     {
+        // Check to see if user is Admin, if not return unautorized message.
+        $user = User::where('username', auth()->user()->username)->first();
+        if ($user == null || ! $user->isAdmin()) {
+            abort(401, 'Not authorized.');
+        }
+
         $rs = RestService::find($rest_service_id);
         $username = auth()->user()->username;
 
@@ -379,6 +487,12 @@ class AdminController extends Controller
 
     public function getUpdateChunkSize($id)
     {
+        // Check to see if user is Admin, if not return unautorized message.
+        $user = User::where('username', auth()->user()->username)->first();
+        if ($user == null || ! $user->isAdmin()) {
+            abort(401, 'Not authorized.');
+        }
+
         $rs = RestService::find($id);
 
         $info = $rs->refreshInfo();
@@ -415,6 +529,12 @@ class AdminController extends Controller
 
     public function getFieldNames($api_version = null)
     {
+        // Check to see if user is Admin, if not return unautorized message.
+        $user = User::where('username', auth()->user()->username)->first();
+        if ($user == null || ! $user->isAdmin()) {
+            abort(401, 'Not authorized.');
+        }
+
         $api_version = $api_version ?? config('ireceptor.default_api_version');
 
         $api_version_list = FieldName::getAPIVersions();
@@ -431,6 +551,12 @@ class AdminController extends Controller
 
     public function queries($nb_months = null)
     {
+        // Check to see if user is Admin, if not return unautorized message.
+        $user = User::where('username', auth()->user()->username)->first();
+        if ($user == null || ! $user->isAdmin()) {
+            abort(401, 'Not authorized.');
+        }
+
         $data = [];
         $data['nb_months'] = $nb_months;
         $data['queries'] = QueryLog::find_gateway_queries($nb_months);
@@ -446,6 +572,12 @@ class AdminController extends Controller
 
     public function queries2($nb_months = null)
     {
+        // Check to see if user is Admin, if not return unautorized message.
+        $user = User::where('username', auth()->user()->username)->first();
+        if ($user == null || ! $user->isAdmin()) {
+            abort(401, 'Not authorized.');
+        }
+
         $data = [];
         $data['nb_months'] = $nb_months;
         $query_list = QueryLog::find_gateway_queries($nb_months);
@@ -463,16 +595,34 @@ class AdminController extends Controller
 
     public function queriesMonths2($n)
     {
+        // Check to see if user is Admin, if not return unautorized message.
+        $user = User::where('username', auth()->user()->username)->first();
+        if ($user == null || ! $user->isAdmin()) {
+            abort(401, 'Not authorized.');
+        }
+
         return $this->queries2($n);
     }
 
     public function queriesMonths($n)
     {
+        // Check to see if user is Admin, if not return unautorized message.
+        $user = User::where('username', auth()->user()->username)->first();
+        if ($user == null || ! $user->isAdmin()) {
+            abort(401, 'Not authorized.');
+        }
+
         return $this->queries($n);
     }
 
     public function query($id)
     {
+        // Check to see if user is Admin, if not return unautorized message.
+        $user = User::where('username', auth()->user()->username)->first();
+        if ($user == null || ! $user->isAdmin()) {
+            abort(401, 'Not authorized.');
+        }
+
         $data = [];
         $data['query_id'] = $id;
         $data['q'] = QueryLog::find($id);
@@ -483,6 +633,12 @@ class AdminController extends Controller
 
     public function downloads()
     {
+        // Check to see if user is Admin, if not return unautorized message.
+        $user = User::where('username', auth()->user()->username)->first();
+        if ($user == null || ! $user->isAdmin()) {
+            abort(401, 'Not authorized.');
+        }
+
         $download_list = Download::orderBy('id', 'desc')->get();
 
         $data = [];
@@ -493,6 +649,12 @@ class AdminController extends Controller
 
     public function jobs()
     {
+        // Check to see if user is Admin, if not return unautorized message.
+        $user = User::where('username', auth()->user()->username)->first();
+        if ($user == null || ! $user->isAdmin()) {
+            abort(401, 'Not authorized.');
+        }
+
         // Get the list of jobs
         $job_list = Job::orderBy('updated_at', 'desc')->get();
         // The job info doesn't have username, just user ID. Add the
@@ -513,6 +675,12 @@ class AdminController extends Controller
 
     public function downloadsMultipleIPAs()
     {
+        // Check to see if user is Admin, if not return unautorized message.
+        $user = User::where('username', auth()->user()->username)->first();
+        if ($user == null || ! $user->isAdmin()) {
+            abort(401, 'Not authorized.');
+        }
+
         $download_list = Download::orderBy('id', 'desc')->get();
 
         $download_list_filtered = [];
